refactor: event type api <-> internal transformers
