+++ packages/lib/event-types/transformers/api-to-internal/api-request.spec.ts
@@ -2,24 +2,24 @@ import { describe, expect, it } from "vitest";
 
 import { FrequencyInput } from "@calcom/platform-enums/monorepo";
 import type {
-  BookingField_2024_06_14,
+  InputBookingField_2024_06_14,
   Location_2024_06_14,
   BookingLimitsCount_2024_06_14,
   BookingWindow_2024_06_14,
   Recurrence_2024_06_14,
 } from "@calcom/platform-types";
 
-import type { UserField } from "./api-request";
 import {
-  transformApiEventTypeLocations,
-  transformApiEventTypeBookingFields,
-  transformSelectOptions,
-  transformApiEventTypeIntervalLimits,
-  transformApiEventTypeFutureBookingLimits,
-  transformApiEventTypeRecurrence,
-} from "./api-request";
+  transformLocationsApiToInternal,
+  transformBookingFieldsApiToInternal,
+  transformSelectOptionsApiToInternal,
+  transformIntervalLimitsApiToInternal,
+  transformFutureBookingLimitsApiToInternal,
+  transformRecurrenceApiToInternal,
+} from "./index";
+import type { CustomField } from "./index";
 
-describe("transformApiEventTypeLocations", () => {
+describe("transformLocationsApiToInternal", () => {
   it("should transform address", () => {
     const input: Location_2024_06_14[] = [
       {
@@ -31,7 +31,7 @@ describe("transformApiEventTypeLocations", () => {
 
     const expectedOutput = [{ type: "inPerson", address: "London road 10-1", displayLocationPublicly: true }];
 
-    const result = transformApiEventTypeLocations(input);
+    const result = transformLocationsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -49,7 +49,7 @@ describe("transformApiEventTypeLocations", () => {
       { type: "link", link: "https://customvideo.com/join/123456", displayLocationPublicly: true },
     ];
 
-    const result = transformApiEventTypeLocations(input);
+    const result = transformLocationsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -64,7 +64,7 @@ describe("transformApiEventTypeLocations", () => {
 
     const expectedOutput = [{ type: "integrations:daily" }];
 
-    const result = transformApiEventTypeLocations(input);
+    const result = transformLocationsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -82,85 +82,15 @@ describe("transformApiEventTypeLocations", () => {
       { type: "userPhone", hostPhoneNumber: "+37120993151", displayLocationPublicly: true },
     ];
 
-    const result = transformApiEventTypeLocations(input);
+    const result = transformLocationsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 });
 
-describe("transformApiEventTypeBookingFields", () => {
-  it("should transform name field", () => {
-    const bookingField: BookingField_2024_06_14 = {
-      type: "name",
-      slug: "name",
-      label: "Your name",
-      required: true,
-      placeholder: "alice",
-    };
-
-    const input: BookingField_2024_06_14[] = [bookingField];
-
-    const expectedOutput: UserField[] = [
-      {
-        name: bookingField.slug,
-        type: bookingField.type,
-        label: bookingField.label,
-        sources: [
-          {
-            id: "user",
-            type: "user",
-            label: "User",
-            fieldRequired: true,
-          },
-        ],
-        editable: "user",
-        required: bookingField.required,
-        placeholder: bookingField.placeholder,
-      },
-    ];
-
-    const result = transformApiEventTypeBookingFields(input);
-
-    expect(result).toEqual(expectedOutput);
-  });
-
-  it("should transform email field", () => {
-    const bookingField: BookingField_2024_06_14 = {
-      type: "email",
-      slug: "email",
-      label: "Your email",
-      required: true,
-      placeholder: "example@example.com",
-    };
-
-    const input: BookingField_2024_06_14[] = [bookingField];
-
-    const expectedOutput: UserField[] = [
-      {
-        name: bookingField.slug,
-        type: bookingField.type,
-        label: bookingField.label,
-        sources: [
-          {
-            id: "user",
-            type: "user",
-            label: "User",
-            fieldRequired: true,
-          },
-        ],
-        editable: "user",
-        required: bookingField.required,
-        placeholder: bookingField.placeholder,
-      },
-    ];
-
-    const result = transformApiEventTypeBookingFields(input);
-
-    expect(result).toEqual(expectedOutput);
-  });
-
+describe("transformBookingFieldsApiToInternal", () => {
   it("should transform phone field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "phone",
       slug: "phone",
       label: "Your phone number",
@@ -168,9 +98,9 @@ describe("transformApiEventTypeBookingFields", () => {
       placeholder: "123456789",
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -189,13 +119,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform address field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "address",
       slug: "address",
       label: "Your address",
@@ -203,9 +133,9 @@ describe("transformApiEventTypeBookingFields", () => {
       placeholder: "1234 Main St",
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -224,13 +154,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform text field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "text",
       slug: "text",
       label: "Your text",
@@ -238,9 +168,9 @@ describe("transformApiEventTypeBookingFields", () => {
       placeholder: "Enter your text",
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -259,13 +189,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform number field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "number",
       slug: "number",
       label: "Your number",
@@ -273,9 +203,9 @@ describe("transformApiEventTypeBookingFields", () => {
       placeholder: "100",
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -294,13 +224,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform textarea field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "textarea",
       slug: "textarea",
       label: "Your detailed information",
@@ -308,9 +238,9 @@ describe("transformApiEventTypeBookingFields", () => {
       placeholder: "Detailed description here...",
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -329,13 +259,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform select field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "select",
       slug: "select",
       label: "Your selection",
@@ -344,9 +274,9 @@ describe("transformApiEventTypeBookingFields", () => {
       options: ["Option 1", "Option 2"],
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -362,17 +292,17 @@ describe("transformApiEventTypeBookingFields", () => {
         editable: "user",
         required: bookingField.required,
         placeholder: bookingField.placeholder,
-        options: transformSelectOptions(bookingField.options),
+        options: transformSelectOptionsApiToInternal(bookingField.options),
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform multiselect field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "multiselect",
       slug: "multiselect",
       label: "Your multiple selections",
@@ -380,9 +310,9 @@ describe("transformApiEventTypeBookingFields", () => {
       options: ["Option 1", "Option 2"],
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -398,17 +328,17 @@ describe("transformApiEventTypeBookingFields", () => {
         editable: "user",
         required: bookingField.required,
         placeholder: "",
-        options: transformSelectOptions(bookingField.options),
+        options: transformSelectOptionsApiToInternal(bookingField.options),
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform multiemail field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "multiemail",
       slug: "multiemail",
       label: "Your multiple emails",
@@ -416,9 +346,9 @@ describe("transformApiEventTypeBookingFields", () => {
       placeholder: "example@example.com",
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -437,13 +367,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform checkbox field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "checkbox",
       slug: "checkbox",
       label: "Your checkboxes",
@@ -451,9 +381,9 @@ describe("transformApiEventTypeBookingFields", () => {
       options: ["Checkbox 1", "Checkbox 2"],
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -469,17 +399,17 @@ describe("transformApiEventTypeBookingFields", () => {
         editable: "user",
         required: bookingField.required,
         placeholder: "",
-        options: transformSelectOptions(bookingField.options),
+        options: transformSelectOptionsApiToInternal(bookingField.options),
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform radio field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "radio",
       slug: "radio",
       label: "Your radio buttons",
@@ -487,9 +417,9 @@ describe("transformApiEventTypeBookingFields", () => {
       options: ["Radio 1", "Radio 2"],
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -505,26 +435,26 @@ describe("transformApiEventTypeBookingFields", () => {
         editable: "user",
         required: bookingField.required,
         placeholder: "",
-        options: transformSelectOptions(bookingField.options),
+        options: transformSelectOptionsApiToInternal(bookingField.options),
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 
   it("should transform boolean field", () => {
-    const bookingField: BookingField_2024_06_14 = {
+    const bookingField: InputBookingField_2024_06_14 = {
       type: "boolean",
       slug: "boolean",
       label: "Agree to terms?",
       required: true,
     };
 
-    const input: BookingField_2024_06_14[] = [bookingField];
+    const input: InputBookingField_2024_06_14[] = [bookingField];
 
-    const expectedOutput: UserField[] = [
+    const expectedOutput: CustomField[] = [
       {
         name: bookingField.slug,
         type: bookingField.type,
@@ -543,13 +473,13 @@ describe("transformApiEventTypeBookingFields", () => {
       },
     ];
 
-    const result = transformApiEventTypeBookingFields(input);
+    const result = transformBookingFieldsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 });
 
-describe("transformApiEventTypeIntervalLimits", () => {
+describe("transformIntervalLimitsApiToInternal", () => {
   it("should transform booking limits count or booking limits duration", () => {
     const input: BookingLimitsCount_2024_06_14 = {
       day: 2,
@@ -564,13 +494,13 @@ describe("transformApiEventTypeIntervalLimits", () => {
       PER_MONTH: 22,
       PER_YEAR: 33,
     };
-    const result = transformApiEventTypeIntervalLimits(input);
+    const result = transformIntervalLimitsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 });
 
-describe("transformApiEventTypeFutureBookingLimits", () => {
+describe("transformFutureBookingLimitsApiToInternal", () => {
   it("should transform range type", () => {
     const input: BookingWindow_2024_06_14 = {
       type: "range",
@@ -583,7 +513,7 @@ describe("transformApiEventTypeFutureBookingLimits", () => {
       periodEndDate: new Date("2024-08-28"),
     };
 
-    const result = transformApiEventTypeFutureBookingLimits(input);
+    const result = transformFutureBookingLimitsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -600,7 +530,7 @@ describe("transformApiEventTypeFutureBookingLimits", () => {
       periodCountCalendarDays: true,
     };
 
-    const result = transformApiEventTypeFutureBookingLimits(input);
+    const result = transformFutureBookingLimitsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -617,7 +547,7 @@ describe("transformApiEventTypeFutureBookingLimits", () => {
       periodCountCalendarDays: true,
     };
 
-    const result = transformApiEventTypeFutureBookingLimits(input);
+    const result = transformFutureBookingLimitsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -634,7 +564,7 @@ describe("transformApiEventTypeFutureBookingLimits", () => {
       periodCountCalendarDays: false,
     };
 
-    const result = transformApiEventTypeFutureBookingLimits(input);
+    const result = transformFutureBookingLimitsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
@@ -651,13 +581,13 @@ describe("transformApiEventTypeFutureBookingLimits", () => {
       periodCountCalendarDays: false,
     };
 
-    const result = transformApiEventTypeFutureBookingLimits(input);
+    const result = transformFutureBookingLimitsApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
 });
 
-describe("transformApiEventTypeRecurrence", () => {
+describe("transformRecurrenceApiToInternal", () => {
   it("should transform recurrence", () => {
     const input: Recurrence_2024_06_14 = {
       frequency: FrequencyInput.weekly,
@@ -669,7 +599,7 @@ describe("transformApiEventTypeRecurrence", () => {
       count: 10,
       freq: 2,
     };
-    const result = transformApiEventTypeRecurrence(input);
+    const result = transformRecurrenceApiToInternal(input);
 
     expect(result).toEqual(expectedOutput);
   });
